fix comment in mysql 8.0
fix comment in mysql 8.0
where is: https://dev.mysql.com/doc/refman/8.0/en/comments.html
"""
From a --  sequence to the end of the line. In MySQL, the --  (double-dash) comment style requires the second dash to be followed by at least one whitespace or control character (such as a space, tab, newline, and so on). This syntax differs slightly from standard SQL comment syntax, as discussed in Section 1.7.2.4, “'--' as the Start of a Comment”.
"""